+++ Horii_HHT/app/src/main/java/com/example/horii_hht/nyuka/Nyuka03_Barread.kt
@@ -4,10 +4,14 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
+import android.content.SharedPreferences
 import android.graphics.Color
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import android.text.Editable
 import android.text.TextWatcher
+import android.util.Log
 import android.view.KeyEvent
 import android.view.View
 import android.widget.Button
@@ -16,6 +20,7 @@ import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
+import androidx.preference.PreferenceManager
 import androidx.room.Room
 import com.cipherlab.barcode.GeneralString
 import com.cipherlab.barcode.ReaderManager
@@ -28,6 +33,9 @@ import com.example.horii_hht.setting.ScreenStateReceiver
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
 import java.util.Timer
 import java.util.TimerTask
 
@@ -47,10 +55,22 @@ class Nyuka03_Barread: AppCompatActivity() {
     private var items: List<Item> = mutableListOf()
 
     private lateinit var screenReceiver: ScreenStateReceiver
+    private val handler = Handler(Looper.getMainLooper()) // ハンドラ
+    private val checkInterval: Long = 10000 // 10秒ごとにチェック
+
+    private val checkRunnable = object : Runnable { // チェック用のRunnable
+        override fun run() {
+            resetDatabaseAndShowDialog() // データベースをリセットしてダイアログ表示
+            handler.postDelayed(this, checkInterval) // 10秒後に再度実行
+        }
+    }
 
     override  fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.nyuka03)
+
+        // 定期的にresetDatabaseAndShowDialogを呼び出す
+        handler.post(checkRunnable)
 
         // ScreenStateReceiverの初期化と登録
         screenReceiver = ScreenStateReceiver()
@@ -366,6 +386,8 @@ class Nyuka03_Barread: AppCompatActivity() {
     // Activity破棄される時に呼び出されるライフサイクルメソッド
     override fun onDestroy() {
         super.onDestroy()
+        // ハンドラのコールバックを削除
+        handler.removeCallbacks(checkRunnable)
         // BroadcastReceiverの解除
         unregisterReceiver(scanDataReceiver)
         // ReaderManagerの解放
@@ -398,21 +420,55 @@ class Nyuka03_Barread: AppCompatActivity() {
             .show()
     }
 
-    //指定した時間分スクリーンオフにしてたら起動するメソッド
+    //指定した時間分放置してたら起動するメソッド
     fun resetDatabaseAndShowDialog() {
-        val lockTime = getLockTimeFromPreferences()
         lifecycleScope.launch(Dispatchers.IO) {
-            dao.deleteAllItems()
-            withContext(Dispatchers.Main) {
-                AlertDialog.Builder(this@Nyuka03_Barread)
-                    .setTitle("注意")
-                    .setMessage("全ての作業を取り消しました。メインメニューに戻ります。")
-                    .setPositiveButton("OK") { _, _ ->
-                        val intent = Intent(this@Nyuka03_Barread, Main_Menu::class.java)
-                        startActivity(intent)
-                        finish()
+            // データベースから最大時間を取得
+            val maxTimes = dao.getMaxTimes()
+            // 取得した最大時間をログに出力
+            Log.d("Nyuka04_Num", "maxTimes: $maxTimes")
+            // 日時フォーマットの設定
+            val dateFormat = SimpleDateFormat("yyyy/MM/dd HH:mm:ss", Locale.getDefault())
+            // maxKenpinTimeが空でない場合に解析
+            val maxKenpinDate = maxTimes?.maxKenpinTime?.takeIf { it.isNotEmpty() }?.let { dateFormat.parse(it) }
+            // maxQRTimeが空でない場合に解析
+            val maxQRDate = maxTimes?.maxQRTime?.takeIf { it.isNotEmpty() }?.let { dateFormat.parse(it) }
+            // maxKenpinDateとmaxQRDateのうち、より直近の時間の方を取得
+            val maxDate = when {
+                maxKenpinDate != null && maxQRDate != null -> maxOf(maxKenpinDate, maxQRDate)
+                maxKenpinDate != null -> maxKenpinDate
+                maxQRDate != null -> maxQRDate
+                else -> null
+            }
+
+            // maxDateがnullでない場合に処理を実行
+            if (maxDate != null) {
+                // 現在の日時を取得
+                val currentDate = Date()
+
+                // 設定した時間を取得
+                val sharedPreferences: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(this@Nyuka03_Barread)
+                val lockTimeMinutes = sharedPreferences.getString("lock_time", "5")?.toLongOrNull() ?: 5
+                val lockTimeMillis = lockTimeMinutes * 60 * 1000
+
+                // 現在の日時と最大時間の差分が設定した時間を超えている場合
+                if (currentDate.time - maxDate.time >= lockTimeMillis) {
+                    // データベースの全アイテムを削除
+                    dao.deleteAllItems()
+                    // メインスレッドでダイアログを表示
+                    withContext(Dispatchers.Main) {
+                        AlertDialog.Builder(this@Nyuka03_Barread)
+                            .setTitle("注意")
+                            .setMessage("経過時間$lockTimeMillis 分。全ての作業を取り消しました。メインメニューに戻ります。")
+                            .setPositiveButton("OK") { _, _ ->
+                                // メインメニューに遷移
+                                val intent = Intent(this@Nyuka03_Barread, Main_Menu::class.java)
+                                startActivity(intent)
+                                finish()
+                            }
+                            .show()
                     }
-                    .show()
+                }
             }
         }
     }
